Replacing `grid.npy` and `lattice_vectors.npy` with `lattice.xyz`.
- The loading in scba needed to change.
