--- examples/w90/carbon-nanotube/inputs/grid.npy
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:66f976b634732a38077e1db5aa69f72ed0c9ec4d4a4b55a051c1ef96a7d162b9
-size 18560
