--- examples/w90/carbon-nanotube/inputs/lattice_vectors.npy
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:69b7a16de6cf5cbc1c659959f38e27ec6ae269a9f74cbba70870c647491dcb1d
-size 200
