--- examples/w90/carbon-nanotube/inputs/wannier_centers.npy
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:929a393bee68372630fd61ac349e9dd1b312be04dc0da89c0f24360df35d96f1
-size 896
